+++ src/uint.rs
@@ -435,232 +435,6 @@ impl_mul! {
     (U4096, 4096)
 }
 
-#[cfg(feature = "cross-size")]
-impl_mul_cross_sizes! {
-    (U64, 64),
-    (
-    (U128, 128),
-    (U256, 256),
-    (U384, 384),
-    (U512, 512),
-    (U640, 640),
-    (U768, 768),
-    (U896, 896),
-    (U1024, 1024),
-    (U1280, 1280),
-    (U1536, 1536),
-    (U1792, 1792),
-    (U2048, 2048),
-    (U3072, 3072),
-    (U3584, 3584),
-    (U4096, 4096),
-    (U6144, 6144)
-    )
-}
-
-#[cfg(feature = "cross-size")]
-impl_mul_cross_sizes! {
-    (U128, 128),
-    (
-        (U256, 256),
-        (U384, 384),
-        (U512, 512),
-        (U640, 640),
-        (U768, 768),
-        (U896, 896),
-        (U1024, 1024),
-        (U1280, 1280),
-        (U1536, 1536),
-        (U1792, 1792),
-        (U2048, 2048),
-        (U3072, 3072),
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144)
-    )
-}
-
-#[cfg(feature = "cross-size")]
-impl_mul_cross_sizes! {
-    (U256, 256),
-    (
-        (U384, 384),
-        (U512, 512),
-        (U640, 640),
-        (U768, 768),
-        (U896, 896),
-        (U1024, 1024),
-        (U1280, 1280),
-        (U1536, 1536),
-        (U1792, 1792),
-        (U2048, 2048),
-        (U3072, 3072),
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144)
-    )
-}
-
-#[cfg(feature = "cross-size")]
-impl_mul_cross_sizes! {
-    (U384, 384),
-    (
-        (U512, 512),
-        (U640, 640),
-        (U768, 768),
-        (U896, 896),
-        (U1024, 1024),
-        (U1280, 1280),
-        (U1536, 1536),
-        (U1792, 1792),
-        (U2048, 2048),
-        (U3072, 3072),
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144)
-    )
-}
-
-#[cfg(feature = "cross-size")]
-impl_mul_cross_sizes! {
-    (U512, 512),
-    (
-        (U640, 640),
-        (U768, 768),
-        (U896, 896),
-        (U1024, 1024),
-        (U1280, 1280),
-        (U1536, 1536),
-        (U1792, 1792),
-        (U2048, 2048),
-        (U3072, 3072),
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144)
-    )
-}
-
-#[cfg(feature = "cross-size")]
-impl_mul_cross_sizes! {
-    (U640, 640),
-    (
-        (U768, 768),
-        (U896, 896),
-        (U1024, 1024),
-        (U1280, 1280),
-        (U1536, 1536),
-        (U1792, 1792),
-        (U2048, 2048),
-        (U3072, 3072),
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144)
-    )
-}
-
-#[cfg(feature = "cross-size")]
-impl_mul_cross_sizes! {
-    (U768, 768),
-    (
-        (U896, 896),
-        (U1024, 1024),
-        (U1280, 1280),
-        (U1536, 1536),
-        (U1792, 1792),
-        (U2048, 2048),
-        (U3072, 3072),
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144)
-    )
-}
-
-#[cfg(feature = "cross-size")]
-impl_mul_cross_sizes! {
-    (U896, 896),
-    (
-        (U1024, 1024),
-        (U1280, 1280),
-        (U1536, 1536),
-        (U1792, 1792),
-        (U2048, 2048),
-        (U3072, 3072),
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144)
-    )
-}
-
-#[cfg(feature = "cross-size")]
-impl_mul_cross_sizes! {
-    (U1024, 1024),
-    (
-        (U1280, 1280),
-        (U1536, 1536),
-        (U1792, 1792),
-        (U2048, 2048),
-        (U3072, 3072),
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144)
-    )
-}
-
-#[cfg(feature = "cross-size")]
-impl_mul_cross_sizes! {
-    (U1280, 1280),
-    (
-        (U1536, 1536),
-        (U1792, 1792),
-        (U2048, 2048),
-        (U3072, 3072),
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144)
-    )
-}
-
-#[cfg(feature = "cross-size")]
-impl_mul_cross_sizes! {
-    (U1536, 1536),
-    (
-        (U1792, 1792),
-        (U2048, 2048),
-        (U3072, 3072),
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144)
-    )
-}
-
-#[cfg(feature = "cross-size")]
-impl_mul_cross_sizes! {
-    (U2048, 2048),
-    (
-        (U3072, 3072),
-        (U3584, 3584),
-        (U4096, 4096),
-        (U6144, 6144)
-    )
-}
-
-#[cfg(feature = "cross-size")]
-impl_mul_cross_sizes! {
-    (U3072, 3072),
-    (
-        (U3584, 3584),
-        (U4096, 4096)
-    )
-}
-
-#[cfg(feature = "cross-size")]
-impl_mul_cross_sizes! {
-    (U3584, 3584),
-    (
-        (U4096, 4096)
-    )
-}
-
 #[cfg(test)]
 mod tests {
     use crate::{Encoding, U128};
