+++ src/uint/concat.rs
@@ -66,11 +66,4 @@ mod tests {
         let res: U128 = U64::ONE.square_wide().into();
         assert_eq!(res, U128::ONE);
     }
-
-    #[cfg(feature = "cross-size")]
-    #[test]
-    fn convert_cross_sizes() {
-        let res: U192 = U64::ONE.mul_wide(&U128::ONE).into();
-        assert_eq!(res, U192::ONE);
-    }
 }
