+++ src/uint/extra_trait_impls.rs
@@ -1,1 +1,2 @@
 mod concat;
+mod mul;
