+++ src/uint/extra_trait_impls/mul.rs
@@ -3,77 +3,79 @@ use crate::{
     U448, U512, U576, U6144, U64, U640, U768, U8192, U896,
 };
 
-use crate::{Concat, Limb};
+use core::ops::Mul;
 
-macro_rules! impl_concat_cross_sizes {
+macro_rules! impl_mul_cross_sizes {
     (($first_type:ident, $first_bits:expr), ($(($second_type:ident, $second_bits:expr)),+ $(,)?)) => {
         $(
-            impl Concat<$second_type> for $first_type {
+            impl Mul<$second_type> for $first_type {
                 type Output = Uint<{nlimbs!($first_bits) + nlimbs!($second_bits)}>;
 
-                fn concat(&self, rhs: &$second_type) -> Self::Output {
-                    let mut limbs = [Limb::ZERO; nlimbs!($first_bits) + nlimbs!($second_bits)];
-                    let mut i = 0;
-                    let mut j = 0;
-
-                    while j < nlimbs!($second_bits) {
-                        limbs[i] = rhs.limbs[j];
-                        i += 1;
-                        j += 1;
-                    }
-
-                    j = 0;
-                    while j < nlimbs!($first_bits) {
-                        limbs[i] = self.limbs[j];
-                        i += 1;
-                        j += 1;
-                    }
-
-                    Uint { limbs }
+                fn mul(self, rhs: $second_type) -> Uint<{nlimbs!($first_bits) + nlimbs!($second_bits)}> {
+                    self.mul_wide(&rhs).into()
                 }
             }
 
-            impl From<($first_type, $second_type)> for Uint<{nlimbs!($first_bits) + nlimbs!($second_bits)}> {
-                fn from(nums: ($first_type, $second_type)) -> Uint<{nlimbs!($first_bits) + nlimbs!($second_bits)}> {
-                    <$second_type as Concat<$first_type>>::concat(&nums.1, &nums.0)
+            impl Mul<&$second_type> for $first_type  {
+                type Output = Uint<{nlimbs!($first_bits) + nlimbs!($second_bits)}>;
+
+                fn mul(self, rhs: &$second_type) -> Uint<{nlimbs!($first_bits) + nlimbs!($second_bits)}> {
+                    self.mul_wide(rhs).into()
                 }
             }
 
-            impl Concat<$first_type> for $second_type {
+            impl Mul<$second_type> for &$first_type  {
+                type Output = Uint<{nlimbs!($first_bits) + nlimbs!($second_bits)}>;
+
+                fn mul(self, rhs: $second_type) -> Uint<{nlimbs!($first_bits) + nlimbs!($second_bits)}> {
+                    self.mul_wide(&rhs).into()
+                }
+            }
+
+            impl Mul<&$second_type> for &$first_type {
+                type Output = Uint<{nlimbs!($first_bits) + nlimbs!($second_bits)}>;
+
+                fn mul(self, rhs: &$second_type) -> Uint<{nlimbs!($first_bits) + nlimbs!($second_bits)}> {
+                    self.mul_wide(rhs).into()
+                }
+            }
+
+            impl Mul<$first_type> for $second_type {
                 type Output = Uint<{nlimbs!($second_bits) + nlimbs!($first_bits)}>;
 
-                fn concat(&self, rhs: &$first_type) -> Self::Output {
-                    let mut limbs = [Limb::ZERO; nlimbs!($second_bits) + nlimbs!($first_bits)];
-                    let mut i = 0;
-                    let mut j = 0;
-
-                    while j < nlimbs!($first_bits) {
-                        limbs[i] = rhs.limbs[j];
-                        i += 1;
-                        j += 1;
-                    }
-
-                    j = 0;
-                    while j < nlimbs!($second_bits) {
-                        limbs[i] = self.limbs[j];
-                        i += 1;
-                        j += 1;
-                    }
-
-                    Uint { limbs }
+                fn mul(self, rhs: $first_type) -> Uint<{nlimbs!($second_bits) + nlimbs!($first_bits)}> {
+                    self.mul_wide(&rhs).into()
                 }
             }
 
-            impl From<($second_type, $first_type)> for Uint<{nlimbs!($second_bits) + nlimbs!($first_bits)}> {
-                fn from(nums: ($second_type, $first_type)) -> Uint<{nlimbs!($second_bits) + nlimbs!($first_bits)}> {
-                    <$first_type as Concat<$second_type>>::concat(&nums.1, &nums.0)
+            impl Mul<&$first_type> for $second_type  {
+                type Output = Uint<{nlimbs!($second_bits) + nlimbs!($first_bits)}>;
+
+                fn mul(self, rhs: &$first_type) -> Uint<{nlimbs!($second_bits) + nlimbs!($first_bits)}> {
+                    self.mul_wide(rhs).into()
+                }
+            }
+
+            impl Mul<$first_type> for &$second_type  {
+                type Output = Uint<{nlimbs!($second_bits) + nlimbs!($first_bits)}>;
+
+                fn mul(self, rhs: $first_type) -> Uint<{nlimbs!($second_bits) + nlimbs!($first_bits)}> {
+                    self.mul_wide(&rhs).into()
+                }
+            }
+
+            impl Mul<&$first_type> for &$second_type {
+                type Output = Uint<{nlimbs!($second_bits) + nlimbs!($first_bits)}>;
+
+                fn mul(self, rhs: &$first_type) -> Uint<{nlimbs!($second_bits) + nlimbs!($first_bits)}> {
+                    self.mul_wide(rhs).into()
                 }
             }
         )+
     };
 }
 
-impl_concat_cross_sizes! {
+impl_mul_cross_sizes! {
     (U64, 64),
     (
     (U128, 128),
@@ -100,7 +102,7 @@ impl_concat_cross_sizes! {
     )
 }
 
-impl_concat_cross_sizes! {
+impl_mul_cross_sizes! {
     (U128, 128),
     (
         (U192, 192),
@@ -126,7 +128,7 @@ impl_concat_cross_sizes! {
     )
 }
 
-impl_concat_cross_sizes! {
+impl_mul_cross_sizes! {
     (U192, 192),
     (
         (U256, 256),
@@ -151,7 +153,7 @@ impl_concat_cross_sizes! {
     )
 }
 
-impl_concat_cross_sizes! {
+impl_mul_cross_sizes! {
     (U256, 256),
     (
         (U320, 320),
@@ -175,7 +177,7 @@ impl_concat_cross_sizes! {
     )
 }
 
-impl_concat_cross_sizes! {
+impl_mul_cross_sizes! {
     (U320, 320),
     (
         (U384, 384),
@@ -198,7 +200,7 @@ impl_concat_cross_sizes! {
     )
 }
 
-impl_concat_cross_sizes! {
+impl_mul_cross_sizes! {
     (U384, 384),
     (
         (U448, 448),
@@ -220,7 +222,7 @@ impl_concat_cross_sizes! {
     )
 }
 
-impl_concat_cross_sizes! {
+impl_mul_cross_sizes! {
     (U448, 448),
     (
         (U512, 512),
@@ -241,7 +243,7 @@ impl_concat_cross_sizes! {
     )
 }
 
-impl_concat_cross_sizes! {
+impl_mul_cross_sizes! {
     (U512, 512),
     (
         (U576, 576),
@@ -261,7 +263,7 @@ impl_concat_cross_sizes! {
     )
 }
 
-impl_concat_cross_sizes! {
+impl_mul_cross_sizes! {
     (U576, 576),
     (
         (U640, 640),
@@ -280,7 +282,7 @@ impl_concat_cross_sizes! {
     )
 }
 
-impl_concat_cross_sizes! {
+impl_mul_cross_sizes! {
     (U640, 640),
     (
         (U768, 768),
@@ -298,7 +300,7 @@ impl_concat_cross_sizes! {
     )
 }
 
-impl_concat_cross_sizes! {
+impl_mul_cross_sizes! {
     (U768, 768),
     (
         (U896, 896),
@@ -315,7 +317,7 @@ impl_concat_cross_sizes! {
     )
 }
 
-impl_concat_cross_sizes! {
+impl_mul_cross_sizes! {
     (U896, 896),
     (
         (U1024, 1024),
@@ -331,7 +333,7 @@ impl_concat_cross_sizes! {
     )
 }
 
-impl_concat_cross_sizes! {
+impl_mul_cross_sizes! {
     (U1024, 1024),
     (
         (U1280, 1280),
@@ -346,7 +348,7 @@ impl_concat_cross_sizes! {
     )
 }
 
-impl_concat_cross_sizes! {
+impl_mul_cross_sizes! {
     (U1280, 1280),
     (
         (U1536, 1536),
@@ -360,7 +362,7 @@ impl_concat_cross_sizes! {
     )
 }
 
-impl_concat_cross_sizes! {
+impl_mul_cross_sizes! {
     (U1536, 1536),
     (
         (U1792, 1792),
@@ -373,7 +375,7 @@ impl_concat_cross_sizes! {
     )
 }
 
-impl_concat_cross_sizes! {
+impl_mul_cross_sizes! {
     (U2048, 2048),
     (
         (U3072, 3072),
@@ -384,7 +386,7 @@ impl_concat_cross_sizes! {
     )
 }
 
-impl_concat_cross_sizes! {
+impl_mul_cross_sizes! {
     (U3072, 3072),
     (
         (U3584, 3584),
@@ -394,7 +396,7 @@ impl_concat_cross_sizes! {
     )
 }
 
-impl_concat_cross_sizes! {
+impl_mul_cross_sizes! {
     (U3584, 3584),
     (
         (U4096, 4096),
@@ -403,7 +405,7 @@ impl_concat_cross_sizes! {
     )
 }
 
-impl_concat_cross_sizes! {
+impl_mul_cross_sizes! {
     (U4096, 4096),
     (
         (U6144, 6144),
@@ -411,7 +413,7 @@ impl_concat_cross_sizes! {
     )
 }
 
-impl_concat_cross_sizes! {
+impl_mul_cross_sizes! {
     (U6144, 6144),
     (
         (U8192, 8192)
@@ -420,30 +422,44 @@ impl_concat_cross_sizes! {
 
 #[cfg(test)]
 mod tests {
-    use crate::{Concat, U128, U192, U64};
+    use crate::{CheckedMul, Zero, U128, U256, U384, U64};
 
     #[test]
-    fn concat_other() {
-        let hi = U64::from_u64(0x0011223344556677);
-        let lo = U128::from_be_hex("00112233445566778899aabbccddeeff");
-
-        assert_eq!(
-            <U64 as Concat<U128>>::concat(&hi, &lo),
-            U192::from_be_hex("001122334455667700112233445566778899aabbccddeeff")
-        );
-
-        let hi = U128::from_be_hex("00112233445566778899aabbccddeeff");
-        let lo = U64::from_u64(0x0011223344556677);
-
-        assert_eq!(
-            <U128 as Concat<U64>>::concat(&hi, &lo),
-            U192::from_be_hex("00112233445566778899aabbccddeeff0011223344556677")
-        );
+    fn mul_wide_zero_and_one_cross_sizes() {
+        assert_eq!(U128::ZERO.mul_wide(&U256::ZERO), (U256::ZERO, U128::ZERO));
+        assert_eq!(U128::ZERO.mul_wide(&U256::ONE), (U256::ZERO, U128::ZERO));
+        assert_eq!(U128::ONE.mul_wide(&U256::ZERO), (U256::ZERO, U128::ZERO));
+        assert_eq!(U128::ONE.mul_wide(&U256::ONE), (U256::ONE, U128::ZERO));
     }
 
     #[test]
-    fn convert_cross_sizes() {
-        let res: U192 = U64::ONE.mul_wide(&U128::ONE).into();
-        assert_eq!(res, U192::ONE);
+    fn mul_wide_cross_sizes() {
+        let x = U128::from_be_hex("ffffffffffffffffffffffffffffffff");
+        let y =
+            U256::from_be_hex("0fffffffffffffffffffffafffffffffffffffffffffffffffffffffffffffff");
+        let (lo, hi) = x.mul_wide(&y);
+
+        assert_eq!(
+            lo,
+            U256::from_be_hex("f0000000000000000000004fffffffff00000000000000000000000000000001")
+        );
+
+        assert_eq!(hi, U128::from_be_hex("0fffffffffffffffffffffafffffffff"));
+    }
+
+    #[test]
+    fn mul_cross_sizes() {
+        let x = U128::from_be_hex("ffffffffffffffffffffffffffffffff");
+        let y =
+            U256::from_be_hex("0fffffffffffffffffffffafffffffffffffffffffffffffffffffffffffffff");
+        let xy: U384 = x.mul_wide(&y).into();
+
+        assert_eq!(xy, x * y);
+        assert_eq!(xy, y * x);
+
+        assert_eq!(
+            xy,
+            U384::from_be_hex("0fffffffffffffffffffffaffffffffff0000000000000000000004fffffffff00000000000000000000000000000001")
+        );
     }
 }
