+++ src/uint/mul.rs
@@ -295,81 +295,8 @@ macro_rules! impl_mul {
     };
 }
 
-#[cfg(feature = "cross-size")]
-macro_rules! impl_mul_cross_sizes {
-    (($first_type:ident, $first_bits:expr), ($(($second_type:ident, $second_bits:expr)),+ $(,)?)) => {
-        $(
-            impl Mul<$second_type> for $first_type {
-                type Output = Uint<{nlimbs!($first_bits) + nlimbs!($second_bits)}>;
-
-                fn mul(self, rhs: $second_type) -> Uint<{nlimbs!($first_bits) + nlimbs!($second_bits)}> {
-                    self.mul_wide(&rhs).into()
-                }
-            }
-
-            impl Mul<&$second_type> for $first_type  {
-                type Output = Uint<{nlimbs!($first_bits) + nlimbs!($second_bits)}>;
-
-                fn mul(self, rhs: &$second_type) -> Uint<{nlimbs!($first_bits) + nlimbs!($second_bits)}> {
-                    self.mul_wide(rhs).into()
-                }
-            }
-
-            impl Mul<$second_type> for &$first_type  {
-                type Output = Uint<{nlimbs!($first_bits) + nlimbs!($second_bits)}>;
-
-                fn mul(self, rhs: $second_type) -> Uint<{nlimbs!($first_bits) + nlimbs!($second_bits)}> {
-                    self.mul_wide(&rhs).into()
-                }
-            }
-
-            impl Mul<&$second_type> for &$first_type {
-                type Output = Uint<{nlimbs!($first_bits) + nlimbs!($second_bits)}>;
-
-                fn mul(self, rhs: &$second_type) -> Uint<{nlimbs!($first_bits) + nlimbs!($second_bits)}> {
-                    self.mul_wide(rhs).into()
-                }
-            }
-
-            impl Mul<$first_type> for $second_type {
-                type Output = Uint<{nlimbs!($second_bits) + nlimbs!($first_bits)}>;
-
-                fn mul(self, rhs: $first_type) -> Uint<{nlimbs!($second_bits) + nlimbs!($first_bits)}> {
-                    self.mul_wide(&rhs).into()
-                }
-            }
-
-            impl Mul<&$first_type> for $second_type  {
-                type Output = Uint<{nlimbs!($second_bits) + nlimbs!($first_bits)}>;
-
-                fn mul(self, rhs: &$first_type) -> Uint<{nlimbs!($second_bits) + nlimbs!($first_bits)}> {
-                    self.mul_wide(rhs).into()
-                }
-            }
-
-            impl Mul<$first_type> for &$second_type  {
-                type Output = Uint<{nlimbs!($second_bits) + nlimbs!($first_bits)}>;
-
-                fn mul(self, rhs: $first_type) -> Uint<{nlimbs!($second_bits) + nlimbs!($first_bits)}> {
-                    self.mul_wide(&rhs).into()
-                }
-            }
-
-            impl Mul<&$first_type> for &$second_type {
-                type Output = Uint<{nlimbs!($second_bits) + nlimbs!($first_bits)}>;
-
-                fn mul(self, rhs: &$first_type) -> Uint<{nlimbs!($second_bits) + nlimbs!($first_bits)}> {
-                    self.mul_wide(rhs).into()
-                }
-            }
-        )+
-    };
-}
-
 #[cfg(test)]
 mod tests {
-    #[cfg(feature = "cross-size")]
-    use crate::U384;
     use crate::{CheckedMul, Zero, U128, U256, U64};
 
     #[test]
@@ -378,11 +305,6 @@ mod tests {
         assert_eq!(U64::ZERO.mul_wide(&U64::ONE), (U64::ZERO, U64::ZERO));
         assert_eq!(U64::ONE.mul_wide(&U64::ZERO), (U64::ZERO, U64::ZERO));
         assert_eq!(U64::ONE.mul_wide(&U64::ONE), (U64::ONE, U64::ZERO));
-
-        assert_eq!(U128::ZERO.mul_wide(&U256::ZERO), (U256::ZERO, U128::ZERO));
-        assert_eq!(U128::ZERO.mul_wide(&U256::ONE), (U256::ZERO, U128::ZERO));
-        assert_eq!(U128::ONE.mul_wide(&U256::ZERO), (U256::ZERO, U128::ZERO));
-        assert_eq!(U128::ONE.mul_wide(&U256::ONE), (U256::ONE, U128::ZERO));
     }
 
     #[test]
@@ -443,22 +365,6 @@ mod tests {
         assert_eq!(hi, U256::MAX.wrapping_sub(&U256::ONE));
     }
 
-    #[cfg(feature = "cross-size")]
-    #[test]
-    fn mul_wide_cross_sizes() {
-        let x = U128::from_be_hex("ffffffffffffffffffffffffffffffff");
-        let y =
-            U256::from_be_hex("0fffffffffffffffffffffafffffffffffffffffffffffffffffffffffffffff");
-        let (lo, hi) = x.mul_wide(&y);
-
-        assert_eq!(
-            lo,
-            U256::from_be_hex("f0000000000000000000004fffffffff00000000000000000000000000000001")
-        );
-
-        assert_eq!(hi, U128::from_be_hex("0fffffffffffffffffffffafffffffff"));
-    }
-
     #[test]
     fn mul() {
         let x = U128::from_be_hex("ffffffffffffffffffffffffffffffff");
@@ -470,23 +376,6 @@ mod tests {
         assert_eq!(
             xy,
             U256::from_be_hex("0fffffffffffffffffffffaffffffffef0000000000000000000005000000001")
-        );
-    }
-
-    #[cfg(feature = "cross-size")]
-    #[test]
-    fn mul_cross_sizes() {
-        let x = U128::from_be_hex("ffffffffffffffffffffffffffffffff");
-        let y =
-            U256::from_be_hex("0fffffffffffffffffffffafffffffffffffffffffffffffffffffffffffffff");
-        let xy: U384 = x.mul_wide(&y).into();
-
-        assert_eq!(xy, x * y);
-        assert_eq!(xy, y * x);
-
-        assert_eq!(
-            xy,
-            U384::from_be_hex("0fffffffffffffffffffffaffffffffff0000000000000000000004fffffffff00000000000000000000000000000001")
         );
     }
 }
